+++ src/jobs/QueueSingleton.php
@@ -7,7 +7,7 @@ namespace cooltronicpl\varnishcache\jobs;
  * CDN Cache & Preload to static HTML Helper Plugin with http & htttps
  *
  * @link      https://cooltronic.pl
- * @copyright Copyright (c) 2023 CoolTRONIC.pl sp. z o.o.
+ * @copyright Copyright (c) 2024 CoolTRONIC.pl sp. z o.o.
  * @author    Pawel Potacki
  */
 use craft\helpers\Queue;