+++ src/variables/VarnishCacheClear.php
@@ -5,7 +5,7 @@
  * CDN Cache & Preload to static HTML Helper Plugin with http & htttps
  *
  * @link      https://cooltronic.pl
- * @copyright Copyright (c) 2023 CoolTRONIC.pl sp. z o.o.
+ * @copyright Copyright (c) 2024 CoolTRONIC.pl sp. z o.o.
  * @author    Pawel Potacki
  */
 
@@ -25,11 +25,11 @@ class VarnishCacheClear
 
     public function clearCustomUrlUriTimeout(string $uri, string $url, int $timeout)
     {
-        \Craft::debug('clearCustomUrlUriTimeout single uri: ' . $uri . " , URL: " . $url);
+        \Craft::info('clearCustomUrlUriTimeout single uri: ' . $uri . " , URL: " . $url);
 
         $v = new VarnishCacheService;
         $v->clearCacheCustomTimeout($uri, $url, $timeout);
-        \Craft::debug('clearCustomUrlUriTimeout single URL ended');
+        \Craft::info('clearCustomUrlUriTimeout single URL ended');
         return "";
 
     }